[core] Move GridFilterModel in the models directory
