+++ docs/pages/api-docs/data-grid/grid-api.md
@@ -72,7 +72,7 @@ import { GridApi } from '@material-ui/x-grid';
 | <span class="prop-name">setEditCellProps</span>                                                   | <span class="prop-type">(params: GridEditCellPropsParams) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                      | Sets the input props of the edit cell.                                                                                                                         |
 | <span class="prop-name">setEditCellValue</span>                                                   | <span class="prop-type">(params: GridEditCellValueParams, event?: SyntheticEvent&lt;Element, Event&gt;) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                        | Sets the value of the edit cell.<br />Commonly used inside the edit cell component.                                                                            |
 | <span class="prop-name">setEditRowsModel</span>                                                   | <span class="prop-type">(model: GridEditRowsModel) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                             | Set sthe edit rows model of the grid.                                                                                                                          |
-| <span class="prop-name">setFilterModel</span>                                                     | <span class="prop-type">(model: GridFilterModelState) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                          | Sets the filter model to the one given by `model`.                                                                                                             |
+| <span class="prop-name">setFilterModel</span>                                                     | <span class="prop-type">(model: GridFilterModel) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                               | Sets the filter model to the one given by `model`.                                                                                                             |
 | <span class="prop-name">setPage</span>                                                            | <span class="prop-type">(page: number) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                                         | Sets the displayed page to the value given by `page`.                                                                                                          |
 | <span class="prop-name">setPageSize</span>                                                        | <span class="prop-type">(pageSize: number) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                                     | Sets the number of displayed rows to the value given by `pageSize`.                                                                                            |
 | <span class="prop-name">setRows</span>                                                            | <span class="prop-type">(rows: GridRowData[]) =&gt; void</span>                                                                                                                                                                                                                                                                                                                                                                                                                                                  | Sets a new set of rows.                                                                                                                                        |
