+++ packages/grid/_modules_/grid/hooks/features/filter/useGridFilter.ts
@@ -21,7 +21,8 @@ import { useGridSelector } from '../core/useGridSelector';
 import { useGridState } from '../core/useGridState';
 import { GridPreferencePanelsValue } from '../preferencesPanel/gridPreferencePanelsValue';
 import { sortedGridRowsSelector } from '../sorting/gridSortingSelector';
-import { getInitialGridFilterState, GridFilterModel } from './gridFilterModelState';
+import { getInitialGridFilterState } from './gridFilterModelState';
+import { GridFilterModel } from '../../../models/gridFilterModel';
 import { visibleSortedGridRowsSelector } from './gridFilterSelector';
 import { getInitialVisibleGridRowsState } from './visibleGridRowsState';
