+++ packages/grid/_modules_/grid/models/gridFilterModel.ts
@@ -1,0 +1,6 @@
+import { GridFilterItem, GridLinkOperator } from './gridFilterItem';
+
+export interface GridFilterModel {
+  items: GridFilterItem[];
+  linkOperator?: GridLinkOperator;
+}
